pyproject.toml
[tool.poetry]
name = "python-cognito-cerbos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8,<4"
cerbos = "^0.1.0"
boto3 = "^1.22.5"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
requests = "^2.27.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
